fix make sure setting migration is not run on facility server

+++ packages/shared/src/migrations/1717088025227-addInvoiceSlidingFeeScaleConfigSetting.js
@@ -1,5 +1,6 @@
 /** @typedef {import('sequelize').QueryInterface} QueryInterface */
 import { Op } from 'sequelize';
+import config from 'config';
 
 const DEFAULT_SETTINGS = {
   'invoice.slidingFeeScale': JSON.stringify([
@@ -22,6 +23,7 @@ const DEFAULT_SETTINGS = {
  * @param {QueryInterface} query
  */
 export async function up(query) {
+  if (config.serverFacilityId) return;
   await query.bulkInsert(
     'settings',
     Object.entries(DEFAULT_SETTINGS).map(([key, value]) => ({
@@ -35,6 +37,7 @@ export async function up(query) {
  * @param {QueryInterface} query
  */
 export async function down(query) {
+  if (config.serverFacilityId) return;
   await query.bulkDelete('settings', {
     key: {
       [Op.in]: Object.keys(DEFAULT_SETTINGS),
